requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain-openai
langchain-google-genai
crewai
sqlite3
gunicorn